+++ tests/test_text.py
@@ -151,3 +151,9 @@ def test_text_sanitize():
         "This is a book description with some and elements. It has paragraphs & "
         "special characters. And multiple paragraphs."
     )
+
+
+def test_text_keep(epub_file, monkeypatch):
+    monkeypatch.setattr(idk.sdk_opts, "text_keep", True)
+    meta = idk.code_text(epub_file)
+    assert meta.text.startswith("THE CONTENTS")